+++ src/Menu.h
@@ -27,5 +27,5 @@ public:
   Menu();
   virtual void Render(sf::RenderWindow *window) override;
   virtual void Input(sf::RenderWindow *window) override;
-  virtual void Update(sf::RenderWindow *window, double delta) override;
+  virtual void Update(sf::RenderWindow *window, float delta) override;
 };
